+++ test/protocol/http/header/trailer.rb
@@ -6,7 +6,7 @@
 require "protocol/http/header/trailer"
 
 describe Protocol::HTTP::Header::Trailer do
-	let(:header) {subject.new(description)}
+	let(:header) {subject.parse(description)}
 	
 	with "etag" do
 		it "contains etag header" do
@@ -59,7 +59,7 @@ describe Protocol::HTTP::Header::Trailer do
 	end
 	
 	with "#<<" do
-		let(:header) {subject.new("etag")}
+		let(:header) {subject.parse("etag")}
 		
 		it "can add headers" do
 			header << "content-md5, expires"