+++ test/protocol/http/header/transfer_encoding.rb
@@ -6,7 +6,7 @@
 require "protocol/http/header/transfer_encoding"
 
 describe Protocol::HTTP::Header::TransferEncoding do
-	let(:header) {subject.new(description)}
+	let(:header) {subject.parse(description)}
 	
 	with "chunked" do
 		it "detects chunked encoding" do
@@ -72,6 +72,24 @@ describe Protocol::HTTP::Header::TransferEncoding do
 	with ".trailer?" do
 		it "should be forbidden in trailers" do
 			expect(subject).not.to be(:trailer?)
+		end
+	end
+	
+	with "normalization" do
+		it "normalizes to lowercase when initialized with string" do
+			header = subject.new("GZIP, CHUNKED")
+			expect(header).to be(:include?, "gzip")
+			expect(header).to be(:include?, "chunked")
+			expect(header).not.to be(:include?, "GZIP")
+			expect(header).not.to be(:include?, "CHUNKED")
+		end
+		
+		it "normalizes to lowercase when initialized with array" do
+			header = subject.new(["GZIP", "CHUNKED"])
+			expect(header).to be(:include?, "gzip")
+			expect(header).to be(:include?, "chunked")
+			expect(header).not.to be(:include?, "GZIP")
+			expect(header).not.to be(:include?, "CHUNKED")
 		end
 	end
 end